Add `name` attribute to GLFS object handles
This is convenience feature to show path by which handle was
opened. This name will also be part of tp_repr for the object
which should help with debugging issues in the bindings. In
case object handle is opened by UUID, the name will be None via
attribute and "<UNKNOWN>" via repr.